+++ apps/web/client/src/components/store/editor/engine/chat/context.ts
@@ -183,9 +183,4 @@ export class ChatContext {
     async clearAttachments() {
         this.context = this.context.filter((context) => context.type !== MessageContextType.IMAGE);
     }
-
-    dispose() {
-        this.clear();
-        this.editorEngine = null as any;
-    }
 }
