+++ apps/web/client/src/components/store/editor/engine/chat/conversation/index.ts
@@ -32,7 +32,7 @@ export class ConversationManager {
     }
 
     async getCurrentProjectConversations(project: Project | null) {
-        this.editorEngine.chat.stream.dispose();
+        this.editorEngine.chat.stream.clear();
         if (!project) {
             return;
         }
@@ -90,7 +90,7 @@ export class ConversationManager {
         }
         this.current = new ChatConversationImpl(this.projectId, []);
         this.conversations.push(this.current);
-        this.editorEngine.chat.stream.dispose();
+        this.editorEngine.chat.stream.clear();
         sendAnalytics('start new conversation');
     }
 
@@ -101,7 +101,7 @@ export class ConversationManager {
             return;
         }
         this.current = match;
-        this.editorEngine.chat.stream.dispose();
+        this.editorEngine.chat.stream.clear();
         sendAnalytics('select conversation');
     }
 
@@ -130,7 +130,7 @@ export class ConversationManager {
                 this.conversations.push(this.current);
             }
         }
-        this.editorEngine.chat.stream.dispose();
+        this.editorEngine.chat.stream.clear();
         sendAnalytics('delete conversation');
     }
