+++ apps/web/client/src/components/store/editor/engine/code/diff.ts
@@ -1,7 +1,5 @@
-import type { CodeDiff, CodeDiffRequest } from '@onlook/models/code';
-import { getAstFromContent } from '@onlook/parser';
-import { generateCode } from './helpers';
-import { transformAst } from './transform';
+import type { CodeDiff, CodeDiffRequest, RequestsByPath } from '@onlook/models';
+import { getAstFromContent, getContentFromAst, transformAst } from "@onlook/parse";
 
 export async function getCodeDiffs(requests: CodeDiffRequest[]): Promise<CodeDiff[]> {
     const groupedRequests = await groupRequestsByOid(requests);
@@ -26,10 +24,31 @@ async function groupRequestsByOid(requests: CodeDiffRequest[]): Promise<Requests
 
         let groupedRequest = groupedRequests.get(path);
         if (!groupedRequest) {
-            groupedRequest = { oidToCodeDiff: new Map(), codeBlock };
+            groupedRequest = {
+                oidToCodeDiff: new Map(), c
+                odeBlock
+            };
         }
         groupedRequest.oidToCodeDiff.set(request.oid, request);
         groupedRequests.set(path, groupedRequest);
     }
     return groupedRequests;
+}
+
+async function processGroupedRequests(groupedRequests: RequestsByPath): Promise<CodeDiff[]> {
+    const diffs: CodeDiff[] = [];
+    for (const [path, request] of groupedRequests) {
+        const { oidToCodeDiff, codeBlock } = request;
+        const ast = getAstFromContent(codeBlock);
+
+        if (!ast) {
+            continue;
+        }
+
+        const original = await getContentFromAst(ast);
+        transformAst(ast, oidToCodeDiff);
+        const generated = await getContentFromAst(ast);
+        diffs.push({ original, generated, path });
+    }
+    return diffs;
 }
