+++ apps/web/client/src/components/store/editor/engine/code/index.ts
@@ -9,12 +9,10 @@ import type {
 import { assertNever } from '@onlook/utility';
 import { debounce } from 'lodash';
 import { makeAutoObservable } from 'mobx';
-import { getEditTextRequests, getGroupRequests, getInsertImageRequests, getInsertRequests, getMoveRequests, getRemoveImageRequests, getRemoveRequests, getStyleRequests, getUngroupRequests } from './requests';
+import { getEditTextRequests, getGroupRequests, getInsertImageRequests, getInsertRequests, getMoveRequests, getRemoveImageRequests, getRemoveRequests, getStyleRequests, getUngroupRequests, getWriteCodeRequests } from './requests';
 
 export class CodeManager {
     isExecuting = false;
-    private writeQueue: Action[] = [];
-    private pendingRequests: CodeDiffRequest[] = [];
 
     constructor(
         private editorEngine: EditorEngine,
@@ -23,6 +21,10 @@ export class CodeManager {
         makeAutoObservable(this);
     }
 
+    /**
+     * Group requests by file then by oid
+     */
+
     async write(action: Action) {
         this.writeQueue.push(action);
         if (!this.isExecuting) {
@@ -30,37 +32,6 @@ export class CodeManager {
             this.debouncedProcessQueue();
         }
     }
-
-    private processQueue = async () => {
-        try {
-            // Process all pending actions and collect the requests
-            while (this.writeQueue.length > 0) {
-                const action = this.writeQueue.shift();
-                if (action) {
-                    const requests = await this.collectRequests(action);
-                    if (Array.isArray(requests)) {
-                        this.pendingRequests.push(...requests);
-                    }
-                }
-            }
-
-            // Write all collected requests at once
-            if (this.pendingRequests.length > 0) {
-                await this.writeRequest(this.pendingRequests);
-                this.pendingRequests = [];
-                sendAnalytics('write code');
-            }
-        } finally {
-            this.isExecuting = false;
-            // Check if more actions were added while processing
-            if (this.writeQueue.length > 0) {
-                this.isExecuting = true;
-                this.debouncedProcessQueue();
-            }
-        }
-    };
-
-    private debouncedProcessQueue = debounce(this.processQueue, 300);
 
     private async collectRequests(action: Action): Promise<CodeDiffRequest[]> {
         switch (action.type) {
@@ -83,7 +54,7 @@ export class CodeManager {
             case 'remove-image':
                 return getRemoveImageRequests(action);
             case 'write-code':
-                return [];
+                return await getWriteCodeRequests(action);
             default:
                 assertNever(action);
         }
@@ -102,10 +73,5 @@ export class CodeManager {
     }
 
     clear() {
-        this.writeQueue = [];
-        this.pendingRequests = [];
-        this.isExecuting = false;
-        this.debouncedProcessQueue.cancel();
-        this.editorEngine = null as any;
     }
 }
